Fixed and Added Repositories
Fixed getUserByUsername, added to the topping repo CostLessThan, CostEquals and CostGreaterThan for toppings
