+++ RPAS-1/src/main/java/dev/calapatia/app/Rpas1Application.java
@@ -1,0 +1,13 @@
+package dev.calapatia.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Rpas1Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Rpas1Application.class, args);
+	}
+
+}
