+++ RPAS-1/src/main/java/dev/supremeteam/repositories/TicketRepository.java
@@ -1,0 +1,13 @@
+package dev.supremeteam.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
+
+import dev.supremeteam.entities.Ticket;
+
+@Component
+@Repository
+public interface TicketRepository extends CrudRepository<Ticket,Integer>{
+//Ticket findByUser(String name);
+}
