+++ RPAS-1/src/main/java/dev/supremeteam/repositories/ToppingRepository.java
@@ -1,0 +1,18 @@
+package dev.supremeteam.repositories;
+
+import java.util.List;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
+
+import dev.supremeteam.entities.Topping;
+
+@Component
+@Repository
+public interface ToppingRepository extends CrudRepository<Topping,Integer>{
+	Topping findByToppingName(String toppingName);
+  List<Topping>     findByCostGreaterThan(float cost);
+  List<Topping>     findByCostEquals(float cost);
+  List<Topping>     findByCostLessThan(float cost);
+}
