+++ RPAS-1/src/main/java/dev/supremeteam/repositories/UserRepository.java
@@ -1,0 +1,13 @@
+package dev.supremeteam.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
+
+import dev.supremeteam.entities.User;
+
+@Component
+@Repository
+public interface UserRepository extends CrudRepository<User,Integer>{
+User findByUsername(String name);
+}
