+++ RPAS-1/src/test/java/dev/calapatia/app/Rpas1ApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.calapatia.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Rpas1ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
