--- SSOOS/bin/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-spring.datasource.url=jdbc:mariadb://rizhallrevdb.crcjhxlklz3b.us-east-2.rds.amazonaws.com:3306/project2superslicedb?user=ahmed&password=pass32word
-server.port=9000
