--- SSOOS/src/main/java/dev/calapatia/app/SsoosApplication.java
@@ -1,13 +1,0 @@
-package dev.calapatia.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SsoosApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SsoosApplication.class, args);
-	}
-
-}
