--- SSOOS/src/main/java/dev/supremeteam/repositories/ToppingRepository.java
@@ -1,20 +1,0 @@
-package dev.supremeteam.repositories;
-
-import java.util.*;
-
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
-
-import dev.supremeteam.entities.Topping;
-
-@Component
-@Repository
-public interface ToppingRepository extends CrudRepository<Topping,Integer>{
-Topping findByToppingName(String toppingName);
-
-	public List<Topping> findByCostLessThan(float price);
-	public List<Topping> findByCostGreaterThan(float price);
-
-}
