--- SSOOS/src/main/java/dev/supremeteam/services/UserService.java
@@ -1,9 +1,0 @@
-package dev.supremeteam.services;
-
-import dev.supremeteam.entities.User;
-
-public interface UserService {
-	User createUser(User u);
-	
-	User loginUser(String username, String password);
-}
