--- SSOOS/src/test/java/dev/calapatia/app/SsoosApplicationTests.java
@@ -1,13 +1,0 @@
-package dev.calapatia.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SsoosApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
